moved seed.js to script/seed.js
